pyproject.toml
[tool.poetry]
name = "gfluent"
version = "0.1.12"
description = "A fluent API for Google Cloud Python Client"
authors = ["Zhong Dai <zhongdai.au@gmail.com>"]
license = "MIT"
readme = "README.md"
documentation = "https://gfluent.readthedocs.io/en/latest/#"
homepage = "https://github.com/zhongdai/gfluent/releases"
repository = "https://github.com/zhongdai/gfluent"
keywords = ["GCP", "BigQuery", "GCS", "Fluent"]


[tool.poetry.dependencies]
python = ">=3.6 <3.10"
google-cloud-bigquery = "^2.13.1"
google-cloud-storage = "^1.37.1"
google-api-python-client = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
poetry = "^1.1.5"
Sphinx = "^3.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
